1. 소유주 충전 이력 (페이징 처리 X) 2. 소유주 시간 변경

+++ SharingCharger_V2/Sources/Model/ChargerModel.swift
@@ -93,7 +93,7 @@ struct OwnerChargerSummary: Codable {
     let currentPoint: Int
     let monthlyReserveCount: Int
     let monthlyRechargeCount: Int
-    let monthlyRechargeKwh: Int
+    let monthlyRechargeKwh: Double
     let monthlyChargerErrorCount: Int
     let monthlyCumulativePoint: Int
 }
